// frontend/shared/src/enums/node-type.ts
export enum NodeType {
  Directory = "Directory",
  File = "File",
  Article = "Article",
  ArticleContent = "ArticleContent",
  Chapter = "Chapter",
  Questionnaire = "Questionnaire",
  QuestionnaireQuestion = "QuestionnaireQuestion",
  CodingModelMainDimension = "CodingModelMainDimension",
  CodingModelSubDimension = "CodingModelSubDimension",
  CodingModelManualItem = "CodingModelManualItem",
  CodingModelAutomatedItem = "CodingModelAutomatedItem",
  InterventionGroupType = "InterventionGroupType",
  Intervention = "Intervention",
  InterventionErpTable = "InterventionErpTable",
  InterventionErpTableRow = "InterventionErpTableDataset",
  InterventionErpRowIntervention = "InterventionErpRowIntervention",
  RScript = "RScript",
  Table = "Table",
  OverviewNode = "OverviewNode"
}